html/semantics/forms/event-propagate-disabled-by-parent.html

<!DOCTYPE html>
<title>Form elements mutated to be disabled during propagation to parents</title>
<meta charset=utf8>
<meta name=timeout content=long>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>

<div id=cases>
  <select></select>
  <select>
    <!-- <option> can't be clicked as it doesn't have its own painting area -->
    <option>foo</option>
  </select>
  <fieldset>Text</fieldset>
  <fieldset><span class="target">Span</span></fieldset>
  <button>Text</button>
  <button><span class="target">Span</span></button>
  <textarea></textarea>
  <input type="button">
  <input type="checkbox">
  <input type="color" value="#000000">
  <input type="date">
  <input type="datetime-local">
  <input type="email">
  <input type="file">
  <input type="image">
  <input type="month">
  <input type="number">
  <input type="password">
  <input type="radio">
  <!-- Native click will click the bar -->
  <input type="range" value="0" data-value="0">
  <!-- Native click will click the slider -->
  <input type="range" value="50" data-value="50">
  <input type="reset">
  <input type="search">
  <input type="submit">
  <input type="tel">
  <input type="text">
  <input type="time">
  <input type="url">
  <input type="week">
</div>

<script>
  /**
   * @param {Element} element
   */
  function getEventFiringTarget(element) {
    return element.querySelector(".target") || element;
  }

  function trusted_click(target) {
    // To workaround type=file clicking issue
    // https://github.com/w3c/webdriver/issues/1666
    return new test_driver.Actions()
      .pointerMove(0, 0, { origin: target })
      .pointerDown()
      .pointerUp()
      .send();
  }

  /**
   * @param {HTMLElement} element
   * @param {object} options
   * @param {boolean} options.initialDisabled
   */
  function initialize(element, { initialDisabled }) {
    element.disabled = initialDisabled;
    switch (element.type) {
      case "checkbox":
      case "radio":
        element.checked = false;
        break;
      case "range":
        element.value = element.dataset.value;
        break;
    }
  }

  /**
   * @param {HTMLElement} element
   * @param {string[]} observedEvents
   * @param {object} options
   * @param {boolean} options.initialDisabled
   */
  function assertStatus(element, observedEvents, { initialDisabled }) {
    switch (element.type) {
      case "checkbox":
      case "radio":
        // By legacy-pre-activation behavior
        // https://html.spec.whatwg.org/multipage/input.html#the-input-element:legacy-canceled-activation-behavior
        assert_equals(element.checked, !initialDisabled, `.checked should be ${!initialDisabled}`);
        break;
      case "range":
        if (initialDisabled) {
          assert_equals(element.value, element.dataset.value, ".value should not change");
        } else {
          assert_true(element.valueAsNumber > 0, ".value should be nonzero");
        }
        break;
    }

    const focusable = element.tabIndex !== -1 && !initialDisabled;

    if (focusable) {
      assert_array_equals(observedEvents, ["focus"], "A focus event should be observed");
    } else {
      assert_array_equals(observedEvents, [], "No events should be observed");
    }
    assert_equals(document.activeElement, document.body, "The focus should be at the body");
  }

  /**
   * @param {Event} ev
   */
  function preventDefaultIfNecessary(element, ev) {
    switch (element.type) {
      case "file": // Opens a modal picker
      case "submit": // May cause a navigation if not
        ev.preventDefault();
    }
  }

  /**
   * @param {*} t
   * @param {Element} element
   * @param {object} options
   * @param {boolean} options.initialDisabled
   */
  function setupTest(t, element, { initialDisabled }) {
    const controller = new AbortController();
    t.add_cleanup(() => controller.abort());
    const { signal } = controller;

    /** @type {string[]} */
    const observedEvents = [];

    initialize(element, { initialDisabled });

    for (const event of ["focus", "blur", "submit"]) {
      element.addEventListener(event, t.step_func(event => {
        observedEvents.push(event.type);
      }), { signal });
    }

    cases.addEventListener("click", t.step_func(event => {
      element.disabled = !initialDisabled;
      preventDefaultIfNecessary(element, event);
    }), { signal });

    const target = getEventFiringTarget(element);
    return { target, observedEvents };
  }

  const elements = document.getElementById("cases").children;
  for (const element of elements) {
    promise_test(async t => {
      const { target, observedEvents } = setupTest(t, element, { initialDisabled: false });

      await trusted_click(target);
      await new Promise(resolve => t.step_timeout(resolve, 0));

      assertStatus(element, observedEvents, { initialDisabled: false });
    }, `immediate propagation on ${element.outerHTML} when initially disabled=false`);


    promise_test(async t => {
      const { target, observedEvents } = setupTest(t, element, { initialDisabled: true });

      await trusted_click(target);
      await new Promise(resolve => t.step_timeout(resolve, 0));

      assertStatus(element, observedEvents, { initialDisabled: true });
    }, `immediate propagation on ${element.outerHTML} when initially disabled=true`);
  }
</script>
